pyproject.toml
[project]
name = "story-history"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "jira>=3.10.5",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=9.0.2",
]

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
